vskernels: bring back zimg kernels
A partial revert of d3a474c. These kernels were rewritten to be custom
kernels in order to implement blur, because at the time we didn't have
support for it through zimg, but now we do via resize2 after
https://github.com/Jaded-Encoding-Thaumaturgy/vapoursynth-resize2/commit/8c3a6beab427848af00322dc223e71308bdf9641.

With this, the Bicubics, Bilinear, Lanczos, Point, and (the relevant)
Spline kernels are back to using resize2/zimg. There are still some
custom kernels like generic Spline, Gaussian, Box, etc., which used to
be implemented via fmtconv, but this patch is just focusing on what can
be done in zimg today.