Dither: update docstring, skip dither in more cases (#8)
* should_dither: skip dither for more full-range integer conversions

8 -> 16 is just a single example of a whole class of conversions
that don't need dithering. VapourSynth supports arbitrary integer depths
from 8 to 32, so more conversions of this class are possible.
Any conversion where the target depth (in bits) is an integer multiple
of the source depth is equivalent to a simple multiplication
by an integer composed of repeated blocks of 00...001 in binary,
e.g. 8 -> 24 (x * 0x10101), 16 -> 32 (x * 0x10001), 9 -> 18 (x * 0x201).

* depth: update doc string describing dither_type default