BestSource: fix throwing exception when there is no environment activated
Originally I wanted the log handler to be removed when doing this:
>>> indexer = BestSource()
>>> ...
>>> del indexer

However, this is too complicated to set up without having vspipe or vspreview throwing a harmless exception.
Finally, adding the logger before the plugin call and removing it just after is probably the simplest way to do even if indexing multiple files in one script will result in multiple added/removed loggers.